Styling CSS for the TABS module
One file change tabs.css
The styles add the following visual options.
The tabs and content panel have a 1px purple border.
The active tab has a 6px ::before top border. As the users arrows through the tabs the ::before top border is moved across the tabs until a tab is activated where the border remains fixed. When focus is moved to the content panel the ::before top border on the active tab changes so that when focus is moved back to the active tab there is a visual indicator it is now selected. THe active tab text is bold.

All colours meet the WCAG 21 Success Criterion 1.4.11 Non-text Contrast all state changes have a minimum content ratio of 3.1

Note: @accdc 
Possible issue TABS
In the previous version of AccDc the attribute aria-selected was triggered when the arrow keys moved focus to the previous next tab. In the 4x version only the tabindex is changed from 0-1.
I have created a pull request with visual styling for TABS. Currently because I have targeted the aria-selected to visually show which element is selected this is currently not working.