Mark-up changes
I have updated the mark-up to replace the dynamic list (dl, dt, dd) as discussed in the following GitHub pages.

The W3C Nu Html Checker flags role errors on the ARIA 1.1 Accordion design pattern. #110
https://github.com/w3c/html-aria/issues/110#issuecomment-398888440

Issue 830: Accordion example update to remove DL, DT and DD elements #838
https://github.com/w3c/aria-practices/pull/838/files/9cf9c8658ec22d7523f358306924d1085c6f684d..ccc49d8f4a93e79019ac3c67401b735a6f1aceb0

Accordion Example: Restructure to remove 'dl' element
https://github.com/w3c/aria-practices/issues/815

Changes:
Replace <dl> with <div>
Replace <dt role heading aria-level with native <h3>
Replaced <div id="accordionGroup"> <div className="accordionGroup"> as no reason to use an ID for this as it is for styling only. If there is a reason to hook to this as an ID this can easily be changed.
Styling applied to the <button> to make the hit area cover the entire trigger region.
Removed extra <span>s and SVG and replaced with FontAwesome.
Added className="content-panel" to have a common class to add padding to the panel content. Can also be used as a hook for styling accordion panel content

Styled accordion with blue background focused/hover changes to a lighter blue and the active (expanded) accordion button has an orange top border.